requirements.txt
django-model-utils